Added user input functionality
user input taken via pop-up prompts in main() function. Messages to user accomplished using alert() function to generate pop-ups. This works fine for now but is not ideal for the print() function, since currently each value in the queue is displayed to the user in a separate pop-up. This will be especially confusing if the user decides to enqueue a lot of values. I plan on fixing this so the print() function will only generate one pop-up displaying the entirety of the queue.