Added search function
Add the search function. This function takes a value from user input and checks to see if that value is currently in the queue. If the value is in the queue, an affirmative message along with that values position in the queue (starting at zero) are displayed in a pop-up window.